+++ library/src/main/java/com/otaliastudios/zoom/ZoomApi.kt
@@ -1,10 +1,8 @@
 package com.otaliastudios.zoom;
 
-import android.graphics.Matrix;
-import androidx.annotation.IntDef;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
+import androidx.annotation.IntDef
+import java.lang.annotation.Retention
+import java.lang.annotation.RetentionPolicy
 
 
 /**
